refact: code organization of unpack
