+++ src/bin/nydus-image/unpack/pax.rs
@@ -1,9 +1,8 @@
 use std::{
     collections::HashMap,
     ffi::OsStr,
-    fs::{File, OpenOptions},
     io::{self, Cursor, Error, ErrorKind, Read},
-    iter::{self, from_fn, repeat},
+    iter::{self, repeat},
     os::unix::prelude::{OsStrExt, OsStringExt},
     path::{Path, PathBuf},
     rc::Rc,
@@ -13,196 +12,24 @@ use std::{
 };
 
 use anyhow::{Context, Result};
-use nydus_api::http::LocalFsConfig;
-use nydus_rafs::{
-    metadata::{RafsInode, RafsMode, RafsSuper},
-    RafsIoReader,
-};
+use nydus_rafs::metadata::RafsInode;
 use nydus_utils::compress::{self, Algorithm};
-use storage::{
-    backend::{localfs::LocalFs, BlobBackend, BlobReader},
-    device::{BlobChunkInfo, BlobInfo},
-    utils::alloc_buf,
-};
-use tar::{Builder, EntryType, Header};
+use storage::{backend::BlobReader, device::BlobChunkInfo, utils::alloc_buf};
+use tar::{EntryType, Header};
 
 use crate::core::node::InodeWrapper;
+
+use super::{SectionBuilder, TarSection};
 
 static PAX_SEP1: &[u8; 1] = b" ";
 static PAX_SEP2: &[u8; 1] = b"=";
 static PAX_PREFIX: &[u8; 13] = b"SCHILY.xattr.";
 static PAX_DELIMITER: &[u8; 1] = b"\n";
 
-pub trait Unpacker {
-    fn unpack(&self) -> Result<()>;
-}
-
-///  A unpacker with the ability to convert bootstrap file and blob file to tar
-pub struct OCIUnpacker {
-    bootstrap: PathBuf,
-    blob: Option<PathBuf>,
-    output: PathBuf,
-
-    builder_factory: OCITarBuilderFactory,
-}
-
-impl OCIUnpacker {
-    pub fn new(bootstrap: &str, blob: Option<&str>, output: &str) -> Result<Self> {
-        let bootstrap = PathBuf::from(bootstrap);
-        let output = PathBuf::from(output);
-        let blob = blob.map(PathBuf::from);
-
-        let builder_factory = OCITarBuilderFactory::new();
-
-        Ok(OCIUnpacker {
-            builder_factory,
-            bootstrap,
-            blob,
-            output,
-        })
-    }
-
-    fn load_rafs(&self) -> Result<RafsSuper> {
-        let bootstrap = OpenOptions::new()
-            .read(true)
-            .write(false)
-            .open(&*self.bootstrap)
-            .with_context(|| format!("fail to open bootstrap {:?}", self.bootstrap))?;
-
-        let mut rs = RafsSuper {
-            mode: RafsMode::Direct,
-            validate_digest: false,
-            ..Default::default()
-        };
-
-        rs.load(&mut (Box::new(bootstrap) as RafsIoReader))
-            .with_context(|| format!("fail to load bootstrap {:?}", self.bootstrap))?;
-
-        Ok(rs)
-    }
-
-    /// A lazy iterator of RafsInode in DFS, which travels in preorder.
-    fn iterator<'a>(
-        &'a self,
-        rs: &'a RafsSuper,
-    ) -> Box<impl Iterator<Item = (Arc<dyn RafsInode>, PathBuf)> + 'a> {
-        // A cursor means the next node to be visited at same height in the tree.
-        // It always starts with the first one of level.
-        // A cursor stack is of cursors from root to leaf.
-        let mut cursor_stack = Vec::with_capacity(32);
-
-        cursor_stack.push(self.cursor_of_root(rs));
-
-        let dfs = move || {
-            while !cursor_stack.is_empty() {
-                let mut cursor = cursor_stack.pop().unwrap();
-
-                let (node, path) = match cursor.next() {
-                    None => continue,
-                    Some(point) => {
-                        cursor_stack.push(cursor);
-                        point
-                    }
-                };
-
-                if node.is_dir() {
-                    cursor_stack.push(self.cursor_of_children(node.clone(), &*path))
-                }
-
-                return Some((node, path));
-            }
-
-            None
-        };
-
-        Box::new(from_fn(dfs))
-    }
-
-    fn cursor_of_children(
-        &self,
-        node: Arc<dyn RafsInode>,
-        path: &Path,
-    ) -> Box<dyn Iterator<Item = (Arc<dyn RafsInode>, PathBuf)>> {
-        let base = path.to_path_buf();
-        let mut next_idx = 0..node.get_child_count();
-
-        let visitor = move || {
-            if next_idx.is_empty() {
-                return None;
-            }
-
-            let child = node.get_child_by_index(next_idx.next().unwrap()).unwrap();
-            let child_path = base.join(child.name());
-
-            Some((child, child_path))
-        };
-
-        Box::new(from_fn(visitor))
-    }
-
-    fn cursor_of_root<'a>(
-        &self,
-        rs: &'a RafsSuper,
-    ) -> Box<dyn Iterator<Item = (Arc<dyn RafsInode>, PathBuf)> + 'a> {
-        let mut has_more = true;
-        let visitor = from_fn(move || {
-            if !has_more {
-                return None;
-            }
-            has_more = false;
-
-            let node = rs.get_inode(rs.superblock.root_ino(), false).unwrap();
-            let path = PathBuf::from("/").join(node.name());
-
-            Some((node, path))
-        });
-
-        Box::new(visitor)
-    }
-}
-
-impl Unpacker for OCIUnpacker {
-    fn unpack(&self) -> Result<()> {
-        info!(
-            "oci unpacker, bootstrap file: {:?}, blob file: {:?}, output file: {:?}",
-            self.bootstrap, self.blob, self.output
-        );
-
-        let rafs = self.load_rafs()?;
-
-        let mut builder = self
-            .builder_factory
-            .create(&rafs, self.blob.as_deref(), &self.output)?;
-
-        for (node, path) in self.iterator(&rafs) {
-            if Util::is_root(&path) {
-                continue;
-            }
-            builder.append(&*node, &path)?;
-        }
-
-        Ok(())
-    }
-}
-
-trait TarBuilder {
-    fn append(&mut self, node: &dyn RafsInode, path: &Path) -> Result<()>;
-}
-
-struct TarSection {
-    header: Header,
-    data: Box<dyn Read>,
-}
-
-trait SectionBuilder {
-    fn can_handle(&mut self, inode: &dyn RafsInode, path: &Path) -> bool;
-    fn build(&self, inode: &dyn RafsInode, path: &Path) -> Result<Vec<TarSection>>;
-}
-
-struct OCISocketBuilder {}
+pub struct OCISocketBuilder {}
 
 impl OCISocketBuilder {
-    fn new() -> Self {
+    pub fn new() -> Self {
         OCISocketBuilder {}
     }
 }
@@ -217,13 +44,13 @@ impl SectionBuilder for OCISocketBuilder {
     }
 }
 
-struct OCILinkBuilder {
-    links: HashMap<u64, Box<Path>>,
+pub struct OCILinkBuilder {
+    links: HashMap<u64, PathBuf>,
     pax_link_builder: Rc<PAXLinkBuilder>,
 }
 
 impl OCILinkBuilder {
-    fn new(pax_link_builder: Rc<PAXLinkBuilder>) -> Self {
+    pub fn new(pax_link_builder: Rc<PAXLinkBuilder>) -> Self {
         OCILinkBuilder {
             links: HashMap::new(),
             pax_link_builder,
@@ -239,8 +66,7 @@ impl SectionBuilder for OCILinkBuilder {
 
         let is_appeared = self.links.contains_key(&node.ino());
         if !is_appeared {
-            self.links
-                .insert(node.ino(), path.to_path_buf().into_boxed_path());
+            self.links.insert(node.ino(), path.to_path_buf());
         }
 
         is_appeared
@@ -254,13 +80,17 @@ impl SectionBuilder for OCILinkBuilder {
     }
 }
 
-struct OCIDirBuilder {
+pub struct OCIDirBuilder {
     ext_builder: Rc<PAXExtensionSectionBuilder>,
 }
 
 impl OCIDirBuilder {
-    fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
+    pub fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
         OCIDirBuilder { ext_builder }
+    }
+
+    fn is_root(&self, path: &Path) -> bool {
+        path.is_absolute() && path.file_name().is_none()
     }
 }
 
@@ -270,6 +100,10 @@ impl SectionBuilder for OCIDirBuilder {
     }
 
     fn build(&self, inode: &dyn RafsInode, path: &Path) -> Result<Vec<TarSection>> {
+        if self.is_root(path) {
+            return Ok(Vec::new());
+        }
+
         let mut header = Header::new_ustar();
         header.set_entry_type(EntryType::dir());
         header.set_size(0);
@@ -307,14 +141,14 @@ impl SectionBuilder for OCIDirBuilder {
     }
 }
 
-struct OCIRegBuilder {
+pub struct OCIRegBuilder {
     ext_builder: Rc<PAXExtensionSectionBuilder>,
     reader: Option<Arc<dyn BlobReader>>,
     compressor: Option<Algorithm>,
 }
 
 impl OCIRegBuilder {
-    fn new(
+    pub fn new(
         ext_builder: Rc<PAXExtensionSectionBuilder>,
         reader: Option<Arc<dyn BlobReader>>,
         compressor: Option<Algorithm>,
@@ -388,12 +222,12 @@ impl SectionBuilder for OCIRegBuilder {
     }
 }
 
-struct OCISymlinkBuilder {
+pub struct OCISymlinkBuilder {
     pax_link_builder: Rc<PAXLinkBuilder>,
 }
 
 impl OCISymlinkBuilder {
-    fn new(pax_link_builder: Rc<PAXLinkBuilder>) -> Self {
+    pub fn new(pax_link_builder: Rc<PAXLinkBuilder>) -> Self {
         OCISymlinkBuilder { pax_link_builder }
     }
 }
@@ -411,12 +245,12 @@ impl SectionBuilder for OCISymlinkBuilder {
     }
 }
 
-struct OCIFifoBuilder {
+pub struct OCIFifoBuilder {
     pax_special_builder: Rc<PAXSpecialSectionBuilder>,
 }
 
 impl OCIFifoBuilder {
-    fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
+    pub fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
         OCIFifoBuilder {
             pax_special_builder,
         }
@@ -434,12 +268,12 @@ impl SectionBuilder for OCIFifoBuilder {
     }
 }
 
-struct OCICharBuilder {
+pub struct OCICharBuilder {
     pax_special_builder: Rc<PAXSpecialSectionBuilder>,
 }
 
 impl OCICharBuilder {
-    fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
+    pub fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
         OCICharBuilder {
             pax_special_builder,
         }
@@ -457,12 +291,12 @@ impl SectionBuilder for OCICharBuilder {
     }
 }
 
-struct OCIBlockBuilder {
+pub struct OCIBlockBuilder {
     pax_special_builder: Rc<PAXSpecialSectionBuilder>,
 }
 
 impl OCIBlockBuilder {
-    fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
+    pub fn new(pax_special_builder: Rc<PAXSpecialSectionBuilder>) -> Self {
         OCIBlockBuilder {
             pax_special_builder,
         }
@@ -480,12 +314,12 @@ impl SectionBuilder for OCIBlockBuilder {
     }
 }
 
-struct PAXSpecialSectionBuilder {
+pub struct PAXSpecialSectionBuilder {
     ext_builder: Rc<PAXExtensionSectionBuilder>,
 }
 
 impl PAXSpecialSectionBuilder {
-    fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
+    pub fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
         PAXSpecialSectionBuilder { ext_builder }
     }
 
@@ -546,10 +380,10 @@ struct PAXRecord {
     v: Vec<u8>,
 }
 
-struct PAXExtensionSectionBuilder {}
+pub struct PAXExtensionSectionBuilder {}
 
 impl PAXExtensionSectionBuilder {
-    fn new() -> Self {
+    pub fn new() -> Self {
         PAXExtensionSectionBuilder {}
     }
 
@@ -635,12 +469,12 @@ impl PAXExtensionSectionBuilder {
     }
 }
 
-struct PAXLinkBuilder {
+pub struct PAXLinkBuilder {
     ext_builder: Rc<PAXExtensionSectionBuilder>,
 }
 
 impl PAXLinkBuilder {
-    fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
+    pub fn new(ext_builder: Rc<PAXExtensionSectionBuilder>) -> Self {
         PAXLinkBuilder { ext_builder }
     }
 
@@ -771,13 +605,9 @@ impl PAXUtil {
     }
 }
 
-struct Util {}
+pub struct Util {}
 
 impl Util {
-    fn is_root(path: &Path) -> bool {
-        path.is_absolute() && path.file_name().is_none()
-    }
-
     fn normalize_path(path: &Path) -> Result<PathBuf> {
         fn end_with_slash(p: &Path) -> bool {
             p.as_os_str().as_bytes().last() == Some(&b'/')
@@ -866,164 +696,6 @@ impl Util {
     }
 }
 
-struct OCITarBuilderFactory {}
-
-impl OCITarBuilderFactory {
-    fn new() -> Self {
-        OCITarBuilderFactory {}
-    }
-
-    fn create(
-        &self,
-        meta: &RafsSuper,
-        blob_path: Option<&Path>,
-        output_path: &Path,
-    ) -> Result<Box<dyn TarBuilder>> {
-        let writer = self.create_writer(output_path)?;
-
-        let blob = meta.superblock.get_blob_infos().pop();
-        let builders = self.create_builders(blob, blob_path)?;
-
-        let builder = OCITarBuilder::new(builders, writer);
-
-        Ok(Box::new(builder) as Box<dyn TarBuilder>)
-    }
-
-    fn create_writer(&self, output_path: &Path) -> Result<Builder<File>> {
-        let builder = Builder::new(
-            OpenOptions::new()
-                .write(true)
-                .create(true)
-                .truncate(true)
-                .read(false)
-                .open(output_path)
-                .with_context(|| format!("fail to open output file {:?}", output_path))?,
-        );
-
-        Ok(builder)
-    }
-
-    fn create_builders(
-        &self,
-        blob: Option<Arc<BlobInfo>>,
-        blob_path: Option<&Path>,
-    ) -> Result<Vec<Box<dyn SectionBuilder>>> {
-        // PAX basic builders
-        let ext_builder = Rc::new(PAXExtensionSectionBuilder::new());
-        let link_builder = Rc::new(PAXLinkBuilder::new(ext_builder.clone()));
-        let special_builder = Rc::new(PAXSpecialSectionBuilder::new(ext_builder.clone()));
-
-        // OCI builders
-        let sock_builder = OCISocketBuilder::new();
-        let hard_link_builder = OCILinkBuilder::new(link_builder.clone());
-        let symlink_builder = OCISymlinkBuilder::new(link_builder);
-        let dir_builder = OCIDirBuilder::new(ext_builder);
-        let fifo_builder = OCIFifoBuilder::new(special_builder.clone());
-        let char_builder = OCICharBuilder::new(special_builder.clone());
-        let block_builder = OCIBlockBuilder::new(special_builder);
-        let reg_builder = self.create_reg_builder(blob, blob_path)?;
-
-        // The order counts.
-        let builders = vec![
-            Box::new(sock_builder) as Box<dyn SectionBuilder>,
-            Box::new(hard_link_builder),
-            Box::new(dir_builder),
-            Box::new(reg_builder),
-            Box::new(symlink_builder),
-            Box::new(fifo_builder),
-            Box::new(char_builder),
-            Box::new(block_builder),
-        ];
-
-        Ok(builders)
-    }
-
-    fn create_reg_builder(
-        &self,
-        blob: Option<Arc<BlobInfo>>,
-        blob_path: Option<&Path>,
-    ) -> Result<OCIRegBuilder> {
-        let (reader, compressor) = match blob {
-            None => (None, None),
-            Some(ref blob) => {
-                if blob_path.is_none() {
-                    bail!("miss blob path")
-                }
-
-                let reader = self.create_blob_reader(blob_path.unwrap())?;
-                let compressor = blob.compressor();
-
-                (Some(reader), Some(compressor))
-            }
-        };
-
-        Ok(OCIRegBuilder::new(
-            Rc::new(PAXExtensionSectionBuilder::new()),
-            reader,
-            compressor,
-        ))
-    }
-
-    fn create_blob_reader(&self, blob_path: &Path) -> Result<Arc<dyn BlobReader>> {
-        let config = LocalFsConfig {
-            blob_file: blob_path.to_str().unwrap().to_owned(),
-            readahead: false,
-            readahead_sec: Default::default(),
-            dir: Default::default(),
-            alt_dirs: Default::default(),
-        };
-        let config = serde_json::to_value(config)
-            .with_context(|| format!("fail to create local backend config for {:?}", blob_path))?;
-
-        let backend = LocalFs::new(config, Some("unpacker"))
-            .with_context(|| format!("fail to create local backend for {:?}", blob_path))?;
-
-        let reader = backend
-            .get_reader("")
-            .map_err(|err| anyhow!("fail to get reader, error {:?}", err))?;
-
-        Ok(reader)
-    }
-}
-
-struct OCITarBuilder {
-    writer: Builder<File>,
-    builders: Vec<Box<dyn SectionBuilder>>,
-}
-
-impl OCITarBuilder {
-    fn new(builders: Vec<Box<dyn SectionBuilder>>, writer: Builder<File>) -> Self {
-        Self { builders, writer }
-    }
-}
-
-impl TarBuilder for OCITarBuilder {
-    fn append(&mut self, inode: &dyn RafsInode, path: &Path) -> Result<()> {
-        let mut is_known_type = true;
-
-        for builder in &mut self.builders {
-            // Useless one, just go !!!!!
-            if !builder.can_handle(inode, path) {
-                continue;
-            }
-
-            for sect in builder.build(inode, path)? {
-                self.writer.append(&sect.header, sect.data)?;
-            }
-
-            is_known_type = false;
-
-            break;
-        }
-
-        if is_known_type {
-            bail!("node {:?} can not be unpacked", path)
-        }
-
-        Ok(())
-    }
-}
-
 struct ChunkReader {
     compressor: Algorithm,
     reader: Arc<dyn BlobReader>,
@@ -1085,17 +757,15 @@ impl Read for ChunkReader {
 
         loop {
             if self.is_chunk_empty() {
-                let chunk = self.chunks.next();
-                if chunk.is_none() {
-                    break;
+                match self.chunks.next() {
+                    None => break,
+                    Some(chunk) => self.load_chunk(chunk.as_ref()).map_err(|err| {
+                        Error::new(
+                            ErrorKind::InvalidData,
+                            format!("fail to load chunk, error: {}", err),
+                        )
+                    })?,
                 }
-
-                self.load_chunk(chunk.unwrap().as_ref()).map_err(|err| {
-                    Error::new(
-                        ErrorKind::InvalidData,
-                        format!("fail to load chunk, error: {}", err),
-                    )
-                })?;
             }
 
             size += Read::read(&mut self.chunk, &mut buf[size..])?;
